Refactored Facts.Utils.read_file/1 to now be able to handle splitting the text based on an optional parameter. If set to true it will split the string by the new line character.
Future implementations may switch this to the default behaviour but
it is currently not the case to maintain compatability until all calls
to the function have been updated.

Fact.Host has a stubbed version of read_file so as to reduce the number
explicit function calls with the new parameter added.

Fact.Host is still currently under development and is progressing nicely.